Master (doc) Fix troubleshooting typos
<!--
BEFORE YOU CREATE A PULL REQUEST:

Ensure you have read over [CONTRIBUTING.md](./CONTRIBUTING.md). We provide VERY defined guidance (as such, we strongly adhere to it).

A summary of our expectations:
 - You are not submitting a pull request from your MASTER / MAIN branch.
 - YOUR GIT COMMIT MESSAGE FORMAT IS EXTREMELY IMPORTANT. We have a very defined expectation for this format and are sticklers about it. Really, READ the entire Contributing document. It will save you and us pain.
 - Do not reformat code, it makes it hard to see what has changed. Leave the formatting to us.

THANKS! We appreciate you reading the entire Contributing document and not just scanning through it.

Name your issue appropriately: give it a sentence that reads well enough for anyone seeing this in release notes to know what the issue is.

When writing out the pull request details please ensure you are writing it as
if you were explaining it to somebody else.
Even if you will be working on and resolving the issue yourself.
This helps others to understand the reasons for the pull request and for it to be searchable in future.

Please do not remove any of the headings.
If a heading is not applicable then enter N/A: Why it's not applicable

Make sure you have raised an issue for this pull request before continuing.

Please remove all comments before submitting.
-->

## Description Of Changes
<!-- Enter a description of the pull request changes -->

## Motivation and Context
<!-- Why is this change necessary and under what context is it being done -->

## Testing

* [ ] I have previewed these changes using the [Docker Container](https://github.com/chocolatey/docs/tree/master/.devcontainer) or another method before submitting this pull request.

## Change Types Made
<!-- Tick the boxes for the type of changes that have been made -->

* [ ] Minor documentation fix (typos etc.).
* [ ] Major documentation change (refactoring, reformatting or adding documentation to existing page).
* [ ] New documentation page added.
* [ ] The change I have made should have a video added, and I have raised an issue for this.
    * Issue #

## Change Checklist

* [ ] Requires a change to menu structure (top or left hand side)/
* [ ] Menu structure has been updated

## Related Issue
<!-- Make sure you have raised an issue for this pull request before continuing. -->

Fixes #

<!-- PLEASE REMOVE ALL COMMENTS BEFORE SUBMITTING -->